Refactor LeafletMap to use dynamic import for better performance
